requirements.txt
accelerate>=0.20.0
awscli
bitsandbytes>=0.39.0
boto3>=1.38.2
botocore>=1.29.0
datasets>=2.14.0
einops>=0.7.0
ipykernel>=6.25.0
jupyter>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
optuna>=3.0.0
pandas>=2.0.0
peft>=0.6.0
psutil>=5.9.0
PyYAML>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.12.0
spacy
tokenizers>=0.13.0
torch>=2.0.0
tqdm>=4.65.0
transformers==4.48.0
wandb>=0.15.0
zstandard>=0.23.0
